Events: Add an initial implementation of register_keystroke_listener()
This is relatively hacky, as we need to keep the Vec of Stashes around
to serve as the backing storage for the KeyDefinition objects, but we
never write to it, so we can only hope and pray that the compiler
doesn't do something stupid with it. This could definitely use some
refactoring: in particular, making this kind of array magic possible
with the GArray crate would be super useful!